分页插件

+++ zc-common-spring-boot-starter/src/main/java/com/zhengcheng/common/constant/CommonConstants.java
@@ -34,7 +34,7 @@ public interface CommonConstants {
     /**
      * 系统默认最大单页限制数量
      */
-    Integer DEFAULT_PAGINATION_LIMIT = 100;
+    Long DEFAULT_PAGINATION_LIMIT = 100L;
     /**
      * 锁KEY的前缀
      */
